feat: Add CSRF protection to form submissions
This commit introduces CSRF protection to form submissions by adding
a hidden input field for the CSRF token in the form templates. It also
updates the 'hx-target' attribute in the logout button to target the
newly added 'response' div. The 'templ' library version is updated
which results in changes in the generated template files.