refactor: Remove unnecessary code and improve security
This commit removes unnecessary print statements and commented out code
across various files. It also improves the security by enhancing the
authentication process in the user model. The commit also includes
changes to the admin handlers and main.go file to accommodate these
updates. The changes also include updates to the user model and the
handling of blog posts, including the way they are displayed and created.