feat: Add About Me page and social icons
This commit introduces a new About Me page to the website. It also
includes the addition of social media icons (LinkedIn, GitHub, Twitter)
which are displayed on the About Me page. The social media icons are
interactive and change color on hover. The CSS styles have been updated
to accommodate these changes. The routing has been updated to include
the new About Me page.