feat: Add dummy post creation for empty database
This commit introduces a new function that checks if there are any
posts in the database. If there are no posts, it creates a dummy
private post. This function is called during the application startup.
This ensures that there is always at least one post in the database,
which can be useful for testing and development purposes.